+++ pom.xml
@@ -38,7 +38,7 @@
                             <transformers>
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>de.exercises.unitconverter.JSONLength</mainClass>
+                                    <mainClass>de.exercises.unitconverter.JSONLengthConverter</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
